+++ ErsatzTV.Application/Troubleshooting/TroubleshootingInfo.cs
@@ -1,11 +1,14 @@
 using ErsatzTV.Application.FFmpegProfiles;
 using ErsatzTV.Core.Domain;
+using ErsatzTV.FFmpeg.Capabilities;
 
 namespace ErsatzTV.Application.Troubleshooting;
 
 public record TroubleshootingInfo(
     string Version,
     Dictionary<string, string> Environment,
+    List<CpuModel> Cpus,
+    List<VideoControllerModel> VideoControllers,
     List<HealthCheckResultSummary> Health,
     FFmpegSettingsViewModel FFmpegSettings,
     List<FFmpegProfile> FFmpegProfiles,